requirements.txt
Flask==2.3.2
pytesseract==0.3.10
Pillow==9.3.0
requests==2.28.1
openai==0.27.0
scikit-learn==1.2.2
numpy==1.24.3
